score_run.py: use multithreading to wait for child procs
It is a waste to use multiprocessing to fork processes, each of
which forks and waits for subprocess. We can just spawn threads,
and let each thread fork and wait for subprocess. Python GIL is
not a concern here as the threads are blocked by the subprocesses
they spawned.

A subprocess forked from a multithreaded process only contains the
working thread that forked it: https://stackoverflow.com/a/1074663

By the way, this change eliminated the necessity of creating a
multiprocessing.Manager() server process (this was to work around
a Python bug on macOS; if interested, see the deleted comment in
git-diff).

Time delta on this change:
13 tests on 16-core 3.1 GHz Linux: 1.00 sec => 0.90 sec
16 tests on 8-core 2.5 GHz macOS:  7.04 sec => 6.06 sec